feat: add classic menu print page with responsive design and print functionality
- Updated router link in mi-pagina.component.html to point to the new classic print page.
- Created menu-classic-print.component.html for displaying the menu in a print-friendly format.
- Added styles in menu-classic-print.component.scss for layout and print optimization.
- Implemented menu-classic-print.component.ts to handle data loading, printing, and displaying tenant information.
- Introduced loading spinner and error handling for better user experience.
