+++ package.json
@@ -26,13 +26,11 @@
         "angularx-qrcode": "^20.0.0",
         "canvas-confetti": "^1.9.3",
         "chart.js": "^4.4.2",
-        "html2canvas": "^1.4.1",
-        "jspdf": "^4.0.0",
         "papaparse": "^5.5.3",
         "primeclt": "^0.1.5",
         "primeicons": "^7.0.0",
         "primeng": "^20.4.0",
-        "quill": "^2.0.3",
+        "quill": "^2.0.2",
         "rxjs": "~7.8.2",
         "tailwindcss-primeui": "^0.6.1",
         "tslib": "^2.8.1",
