+++ src/app/pages/mi-pagina/mi-pagina.component.html
@@ -219,7 +219,7 @@
 
         <div class="flex gap-2 flex-wrap">
           <a
-            routerLink="/dashboard/menu-print"
+            routerLink="/dashboard/menu-classic-print"
             class="p-button p-button-primary p-component"
             style="text-decoration: none;"
           >
